Move stacking config to options builder
