+++ packages/grafana-ui/src/options/builder/index.ts
@@ -2,3 +2,4 @@ export * from './axis';
 export * from './hideSeries';
 export * from './legend';
 export * from './tooltip';
+export * from './stacking';
